+++ stroom-index-server/src/main/java/stroom/search/server/ClusterSearchTask.java
@@ -16,46 +16,53 @@
 
 package stroom.search.server;
 
-import stroom.node.shared.Node;
 import stroom.query.api.v2.Query;
+import stroom.query.api.v2.QueryKey;
 import stroom.query.common.v2.CoprocessorSettings;
 import stroom.query.common.v2.CoprocessorSettingsMap.CoprocessorKey;
-import stroom.search.resultsender.NodeResult;
-import stroom.query.api.v2.Query;
 import stroom.security.shared.UserIdentity;
 import stroom.task.cluster.ClusterTask;
+import stroom.util.shared.VoidResult;
 
 import java.util.List;
 import java.util.Map;
 
-public class ClusterSearchTask extends ClusterTask<NodeResult> {
+public class ClusterSearchTask extends ClusterTask<VoidResult> {
     private static final long serialVersionUID = -1305243739417365803L;
 
+    private final QueryKey key;
     private final Query query;
     private final List<Long> shards;
-    private final Node targetNode;
     private final String[] storedFields;
-    private final int resultSendFrequency;
     private final Map<CoprocessorKey, CoprocessorSettings> coprocessorMap;
     private final String dateTimeLocale;
     private final long now;
 
-    public ClusterSearchTask(final UserIdentity userIdentity, final String taskName, final Query query,
-                             final List<Long> shards, final Node targetNode, final String[] storedFields,
-                             final int resultSendFrequency, final Map<CoprocessorKey, CoprocessorSettings> coprocessorMap, final String dateTimeLocale, final long now) {
+    public ClusterSearchTask(final UserIdentity userIdentity,
+                             final String taskName,
+                             final QueryKey key,
+                             final Query query,
+                             final List<Long> shards,
+                             final String[] storedFields,
+                             final Map<CoprocessorKey, CoprocessorSettings> coprocessorMap,
+                             final String dateTimeLocale,
+                             final long now) {
         super(userIdentity, taskName);
+        this.key = key;
         this.query = query;
         this.shards = shards;
-        this.targetNode = targetNode;
         this.storedFields = storedFields;
-        this.resultSendFrequency = resultSendFrequency;
         this.coprocessorMap = coprocessorMap;
         this.dateTimeLocale = dateTimeLocale;
         this.now = now;
     }
 
-    public Node getTargetNode() {
-        return targetNode;
+//    public Node getTargetNode() {
+//        return targetNode;
+//    }
+
+    public QueryKey getKey() {
+        return key;
     }
 
     public Query getQuery() {
@@ -70,9 +77,9 @@ public class ClusterSearchTask extends ClusterTask<NodeResult> {
         return storedFields;
     }
 
-    public int getResultSendFrequency() {
-        return resultSendFrequency;
-    }
+//    public int getResultSendFrequency() {
+//        return resultSendFrequency;
+//    }
 
     public Map<CoprocessorKey, CoprocessorSettings> getCoprocessorMap() {
         return coprocessorMap;